Test in GHA CI, use make for InspectorXSLT for #72
This change integrates the Makefile-based testing harness for each of
the submodules.

For the InspectorXSLT submodule in the schema-gen directory, integrate
the shell script that orchestrates Maven's installation of dependencies
and execution of XML Calabash to run XProc files. The XProc pipeline
files will run the XSpec tests in batch mode.